+++ app/sites/index/controller.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-arcgis-server-services/sites/index/controller';